requirements.txt
python-telegram-bot==20.7
trafilatura
groq
requests
python-dotenv